+++ lib/transform_reduce.cu
@@ -95,10 +95,10 @@ namespace quda
   template void transform_reduce<plus<double>, complex<signed char>, unsigned int, abs_<double, signed char>, identity>(
     QudaFieldLocation, std::vector<double> &, std::vector<complex<signed char> *> const &, unsigned int,
     abs_<double, signed char>, identity);
-  template double transform_reduce<plus<double>, complex<double>, unsigned long, abs_<double, double>>(
-    QudaFieldLocation, complex<double> const *, unsigned long, abs_<double, double>);
-  template double transform_reduce<plus<double>, complex<float>, unsigned long, abs_<double, float>>(
-    QudaFieldLocation, complex<float> const *, unsigned long, abs_<double, float>);
+  template double transform_reduce<plus<double>, complex<double>, unsigned long long, abs_<double, double>>(
+    QudaFieldLocation, complex<double> const *, unsigned long long, abs_<double, double>);
+  template double transform_reduce<plus<double>, complex<float>, unsigned long long, abs_<double, float>>(
+    QudaFieldLocation, complex<float> const *, unsigned long long, abs_<double, float>);
   // square
   template void transform_reduce<plus<double>, complex<double>, unsigned int, square_<double, double>, identity>(
     QudaFieldLocation, std::vector<double> &, std::vector<complex<double> *> const &, unsigned int,
@@ -115,10 +115,10 @@ namespace quda
   template void transform_reduce<plus<double>, complex<short>, unsigned int, square_<double, short>, identity>(
     QudaFieldLocation, std::vector<double> &, std::vector<complex<short> *> const &, unsigned int,
     square_<double, short>, identity);
-  template double transform_reduce<plus<double>, complex<double>, unsigned long, square_<double, double>>(
-    QudaFieldLocation, complex<double> const *, unsigned long, square_<double, double>);
-  template double transform_reduce<plus<double>, complex<float>, unsigned long, square_<double, float>>(
-    QudaFieldLocation, complex<float> const *, unsigned long, square_<double, float>);
+  template double transform_reduce<plus<double>, complex<double>, unsigned long long, square_<double, double>>(
+    QudaFieldLocation, complex<double> const *, unsigned long long, square_<double, double>);
+  template double transform_reduce<plus<double>, complex<float>, unsigned long long, square_<double, float>>(
+    QudaFieldLocation, complex<float> const *, unsigned long long, square_<double, float>);
   template double transform_reduce<plus<double>, complex<float>, unsigned int, square_<double, float>>(
     QudaFieldLocation, complex<float> const *, unsigned int, square_<double, float>);
   template double transform_reduce<plus<double>, complex<short>, unsigned int, square_<double, short>>(
@@ -140,16 +140,16 @@ namespace quda
   template void transform_reduce<maximum<float>, complex<short>, unsigned int, abs_max_<float, short>, identity>(
     QudaFieldLocation, std::vector<float> &, std::vector<complex<short> *> const &, unsigned int,
     abs_max_<float, short>, identity);
-  template double transform_reduce<maximum<double>, complex<double>, unsigned long, abs_max_<double, double>>(
-    QudaFieldLocation, complex<double> const *, unsigned long, abs_max_<double, double>);
+  template double transform_reduce<maximum<double>, complex<double>, unsigned long long, abs_max_<double, double>>(
+    QudaFieldLocation, complex<double> const *, unsigned long long, abs_max_<double, double>);
   template double transform_reduce<maximum<double>, complex<double>, unsigned int, abs_max_<double, double>>(
     QudaFieldLocation, complex<double> const *, unsigned int, abs_max_<double, double>);
-  template float transform_reduce<maximum<float>, complex<float>, unsigned long, abs_max_<float, float>>(
-    QudaFieldLocation, complex<float> const *, unsigned long, abs_max_<float, float>);
+  template float transform_reduce<maximum<float>, complex<float>, unsigned long long, abs_max_<float, float>>(
+    QudaFieldLocation, complex<float> const *, unsigned long long, abs_max_<float, float>);
   template float transform_reduce<maximum<float>, complex<float>, unsigned int, abs_max_<float, float>>(
     QudaFieldLocation, complex<float> const *, unsigned int, abs_max_<float, float>);
-  template float transform_reduce<maximum<float>, complex<short>, unsigned long, abs_max_<float, short>>(
-    QudaFieldLocation, complex<short> const *, unsigned long, abs_max_<float, short>);
+  template float transform_reduce<maximum<float>, complex<short>, unsigned long long, abs_max_<float, short>>(
+    QudaFieldLocation, complex<short> const *, unsigned long long, abs_max_<float, short>);
   template float transform_reduce<maximum<float>, complex<short>, unsigned int, abs_max_<float, short>>(
     QudaFieldLocation, complex<short> const*, unsigned int, abs_max_<float, short>);
   // abs_min
@@ -169,10 +169,10 @@ namespace quda
   transform_reduce<minimum<float>, complex<signed char>, unsigned int, abs_min_<float, signed char>, identity>(
     QudaFieldLocation, std::vector<float> &, std::vector<complex<signed char> *> const &, unsigned int,
     abs_min_<float, signed char>, identity);
-  template double transform_reduce<minimum<double>, complex<double>, unsigned long, abs_min_<double, double>>(
-    QudaFieldLocation, complex<double> const *, unsigned long, abs_min_<double, double>);
-  template float transform_reduce<minimum<float>, complex<float>, unsigned long, abs_min_<float, float>>(
-    QudaFieldLocation, complex<float> const *, unsigned long, abs_min_<float, float>);
+  template double transform_reduce<minimum<double>, complex<double>, unsigned long long, abs_min_<double, double>>(
+    QudaFieldLocation, complex<double> const *, unsigned long long, abs_min_<double, double>);
+  template float transform_reduce<minimum<float>, complex<float>, unsigned long long, abs_min_<float, float>>(
+    QudaFieldLocation, complex<float> const *, unsigned long long, abs_min_<float, float>);
 
   template void quda::transform_reduce<plus<double>, complex<double>, unsigned int, square_<double, double>, milc_mapper>(
     QudaFieldLocation, std::vector<double> &, std::vector<quda::complex<double> *> const &, unsigned int,
